♻️ ui mode 관련 로직을 feature 레이어에서 수행한다
