+++ src/devtools/models/devtools-view-store.ts
@@ -1,18 +1,16 @@
 import { create } from 'zustand';
-import type { TTab, TUiMode } from '../types';
+import type { TTab } from '../types';
 
 const useDevtoolsViewStore = create<TDevtoolsViewState>(set => ({
-  uiMode: 'maximized',
   tab: 'all',
-  setUiMode: uiMode => set({ uiMode }),
+
   setTab: tab => set({ tab }),
 }));
 
 export { useDevtoolsViewStore };
 
 type TDevtoolsViewState = {
-  uiMode: TUiMode;
   tab: TTab;
-  setUiMode: (uiMode: TUiMode) => void;
+
   setTab: (tab: TTab) => void;
 };
