+++ src/devtools/types/index.ts
@@ -1,2 +1,1 @@
 export type TTab = 'all' | 'http' | 'socketio' | 'sse';
-export type TUiMode = 'maximized' | 'minimized' | 'closed';
