+++ src/features/switch-ui-mode/models/ui-mode-store.tsx
@@ -1,0 +1,17 @@
+import { create } from 'zustand';
+
+const useUiModeStore = create<TUiModeState>(set => ({
+  uiMode: 'closed',
+
+  setUiMode: uiMode => set({ uiMode }),
+}));
+
+export { useUiModeStore };
+
+type TUiModeState = {
+  uiMode: TUiMode;
+
+  setUiMode: (uiMode: TUiMode) => void;
+};
+
+export type TUiMode = 'maximized' | 'minimized' | 'closed';
